+++ docs/changelog.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [2.1.10](https://github.com/SSENSE/node-cucumber-testrail-sync/tree/2.1.10) (2016-12-15)
+
+- Add debug flag when running the synchronization script
+
+- Support for TestRail's steps field for test cases
+
 ## [2.1.8](https://github.com/SSENSE/node-cucumber-testrail-sync/tree/2.1.8) (2016-12-09)
 
 - Handle undefined/skipped scenarios when pushing results to TestRail